+++ src/main/java/com/imsweb/algorithms/behavrecode/SeerBehaviorRecodeAlgorithm.java
@@ -3,10 +3,7 @@
  */
 package com.imsweb.algorithms.behavrecode;
 
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 import com.imsweb.algorithms.AbstractAlgorithm;
@@ -20,7 +17,6 @@ import static com.imsweb.algorithms.Algorithms.FIELD_DX_DATE;
 import static com.imsweb.algorithms.Algorithms.FIELD_HIST_O3;
 import static com.imsweb.algorithms.Algorithms.FIELD_PRIMARY_SITE;
 import static com.imsweb.algorithms.Algorithms.FIELD_SEER_BEHAV_RECODE;
-import static com.imsweb.algorithms.Algorithms.FIELD_TUMORS;
 
 public class SeerBehaviorRecodeAlgorithm extends AbstractAlgorithm {
 
@@ -41,16 +37,14 @@ public class SeerBehaviorRecodeAlgorithm extends AbstractAlgorithm {
 
     @Override
     public AlgorithmOutput execute(AlgorithmInput input) {
-        Map<String, Object> outputPatient = new HashMap<>();
-        List<Map<String, Object>> outputTumors = new ArrayList<>();
-        outputPatient.put(FIELD_TUMORS, outputTumors);
+        Map<String, Object> outputPatient = Utils.createPatientOutput();
 
-        for (Map<String, Object> inputTumor : Utils.extractTumors(Utils.extractPatient(input))) {
+        for (Map<String, Object> inputTumor : Utils.extractTumors(input)) {
             String site = (String)inputTumor.get(FIELD_PRIMARY_SITE);
             String hist = (String)inputTumor.get(FIELD_HIST_O3);
             String beh = (String)inputTumor.get(FIELD_BEHAV_O3);
             String dxYear = Utils.extractYear((String)inputTumor.get(FIELD_DX_DATE));
-            outputTumors.add(Collections.singletonMap(FIELD_SEER_BEHAV_RECODE, BehaviorRecodeUtils.computeBehaviorRecode(site, hist, beh, dxYear)));
+            Utils.addTumorOutput(outputPatient, Collections.singletonMap(FIELD_SEER_BEHAV_RECODE, BehaviorRecodeUtils.computeBehaviorRecode(site, hist, beh, dxYear)));
         }
 
         return AlgorithmOutput.of(outputPatient);