+++ src/main/java/com/imsweb/algorithms/iccc/IcccAlgorithm.java
@@ -3,10 +3,8 @@
  */
 package com.imsweb.algorithms.iccc;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 import com.imsweb.algorithms.AbstractAlgorithm;
@@ -20,7 +18,6 @@ import static com.imsweb.algorithms.Algorithms.FIELD_HIST_O3;
 import static com.imsweb.algorithms.Algorithms.FIELD_ICCC;
 import static com.imsweb.algorithms.Algorithms.FIELD_ICCC_EXT;
 import static com.imsweb.algorithms.Algorithms.FIELD_PRIMARY_SITE;
-import static com.imsweb.algorithms.Algorithms.FIELD_TUMORS;
 
 public class IcccAlgorithm extends AbstractAlgorithm {
 
@@ -42,11 +39,9 @@ public class IcccAlgorithm extends AbstractAlgorithm {
 
     @Override
     public AlgorithmOutput execute(AlgorithmInput input) {
-        Map<String, Object> outputPatient = new HashMap<>();
-        List<Map<String, Object>> outputTumors = new ArrayList<>();
-        outputPatient.put(FIELD_TUMORS, outputTumors);
+        Map<String, Object> outputPatient = Utils.createPatientOutput();
 
-        for (Map<String, Object> inputTumor : Utils.extractTumors(Utils.extractPatient(input))) {
+        for (Map<String, Object> inputTumor : Utils.extractTumors(input)) {
             String site = (String)inputTumor.get(FIELD_PRIMARY_SITE);
             String hist = (String)inputTumor.get(FIELD_HIST_O3);
             String beh = (String)inputTumor.get(FIELD_BEHAV_O3);
@@ -57,7 +52,7 @@ public class IcccAlgorithm extends AbstractAlgorithm {
             Map<String, Object> outputTumor = new HashMap<>();
             outputTumor.put(FIELD_ICCC, icccCode);
             outputTumor.put(FIELD_ICCC_EXT, icccExtCode);
-            outputTumors.add(outputTumor);
+            Utils.addTumorOutput(outputPatient, outputTumor);
         }
 
         return AlgorithmOutput.of(outputPatient);