+++ src/main/java/com/imsweb/algorithms/lymphoma/LymphoidNeoplasmRecodeAlgorithm2021.java
@@ -3,10 +3,7 @@
  */
 package com.imsweb.algorithms.lymphoma;
 
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 import com.imsweb.algorithms.AbstractAlgorithm;
@@ -18,7 +15,6 @@ import com.imsweb.algorithms.internal.Utils;
 import static com.imsweb.algorithms.Algorithms.FIELD_HIST_O3;
 import static com.imsweb.algorithms.Algorithms.FIELD_PRIMARY_SITE;
 import static com.imsweb.algorithms.Algorithms.FIELD_SEER_LYMPH_NEO_RECODE_2021;
-import static com.imsweb.algorithms.Algorithms.FIELD_TUMORS;
 import static com.imsweb.algorithms.lymphoma.LymphoidNeoplasmRecodeUtils.ALG_NAME;
 import static com.imsweb.algorithms.lymphoma.LymphoidNeoplasmRecodeUtils.ALG_VERSION_2021;
 
@@ -39,14 +35,12 @@ public class LymphoidNeoplasmRecodeAlgorithm2021 extends AbstractAlgorithm {
 
     @Override
     public AlgorithmOutput execute(AlgorithmInput input) {
-        Map<String, Object> outputPatient = new HashMap<>();
-        List<Map<String, Object>> outputTumors = new ArrayList<>();
-        outputPatient.put(FIELD_TUMORS, outputTumors);
+        Map<String, Object> outputPatient = Utils.createPatientOutput();
 
-        for (Map<String, Object> inputTumor : Utils.extractTumors(Utils.extractPatient(input))) {
+        for (Map<String, Object> inputTumor : Utils.extractTumors(input)) {
             String site = (String)inputTumor.get(FIELD_PRIMARY_SITE);
             String hist = (String)inputTumor.get(FIELD_HIST_O3);
-            outputTumors.add(Collections.singletonMap(FIELD_SEER_LYMPH_NEO_RECODE_2021, LymphoidNeoplasmRecodeUtils.calculateSiteRecode(ALG_VERSION_2021, site, hist)));
+            Utils.addTumorOutput(outputPatient, Collections.singletonMap(FIELD_SEER_LYMPH_NEO_RECODE_2021, LymphoidNeoplasmRecodeUtils.calculateSiteRecode(ALG_VERSION_2021, site, hist)));
         }
 
         return AlgorithmOutput.of(outputPatient);