+++ src/main/java/com/imsweb/algorithms/persistentpoverty/PersistentPovertyAlgorithm.java
@@ -45,9 +45,8 @@ public class PersistentPovertyAlgorithm extends AbstractAlgorithm {
 
     @Override
     public AlgorithmOutput execute(AlgorithmInput input) {
-
         List<Map<String, Object>> outputTumors = new ArrayList<>();
-        for (Map<String, Object> inputTumor : Utils.extractTumors(Utils.extractPatient(input))) {
+        for (Map<String, Object> inputTumor : Utils.extractTumors(input)) {
             StateCountyTractInputDto inputDto = createStateCountyTractInputDto(inputTumor);
             
             inputDto.applyRecodes();