+++ src/main/java/com/imsweb/algorithms/tractestcongressdist/TractEstCongressDistAlgorithm.java
@@ -41,9 +41,8 @@ public class TractEstCongressDistAlgorithm extends AbstractAlgorithm {
 
     @Override
     public AlgorithmOutput execute(AlgorithmInput input) {
-
         List<Map<String, Object>> outputTumors = new ArrayList<>();
-        for (Map<String, Object> inputTumor : Utils.extractTumors(Utils.extractPatient(input))) {
+        for (Map<String, Object> inputTumor : Utils.extractTumors(input)) {
             TractEstCongressDistOutputDto outputDto = TractEstCongressDistUtils.computeTractEstCongressDist(createStateCountyTractInputDto(inputTumor));
 
             Map<String, Object> outputTumor = new HashMap<>();