+++ src/test/java/com/imsweb/algorithms/survival/SurvivalTimeUtilsTest.java
@@ -22,6 +22,7 @@ public class SurvivalTimeUtilsTest {
     @Test
     @SuppressWarnings("java:S5961") // too much complexity
     public void testLogic() {
+        Assert.assertTrue(SurvivalTimeUtils.calculateSurvivalTime(new SurvivalTimeInputPatientDto(), 2020).getSurvivalTimeOutputPatientDtoList().isEmpty());
 
         List<SurvivalTimeInputRecordDto> records = new ArrayList<>();