+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Changes in version 5.8 (not released yet)**
 
 - Added optional parameter to Cause of Death Classification utility methods to run the logic of the algorithm using a different SEER Site Recode version.
+- Harmonized behavior of processing patient with no tumors.
 
 **Changes in version 5.7**